requirements.txt
joblib==1.0.0
numpy==1.19.5
pandas==1.2.1
python-dateutil==2.8.1
pytz==2020.5
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
requests==2.22.0
azure-functions